+++ src/components/theme-provider.tsx
@@ -1,4 +1,5 @@
-import { createContext, useContext, useEffect, useState } from "react";
+import { useEffect, useState } from "react";
+import { ThemeProviderContext } from "./theme-hooks";
 
 type Theme = "dark" | "light" | "system";
 
@@ -17,8 +18,6 @@ const initialState: ThemeProviderState = {
   theme: "system",
   setTheme: () => null,
 };
-
-const ThemeProviderContext = createContext<ThemeProviderState>(initialState);
 
 export function ThemeProvider({
   children,
@@ -62,12 +61,3 @@ export function ThemeProvider({
     </ThemeProviderContext.Provider>
   );
 }
-
-export const useTheme = () => {
-  const context = useContext(ThemeProviderContext);
-
-  if (context === undefined)
-    throw new Error("useTheme must be used within a ThemeProvider");
-
-  return context;
-};
