+++ src/lib/i18n.tsx
@@ -1,4 +1,8 @@
-import React, { createContext, useContext, useMemo, useState } from "react";
+import React, { useMemo, useState } from "react";
+import { I18nContext } from "./i18n-hooks";
+
+// Re-export the hook for backward compatibility
+export { useI18n } from "./i18n-hooks";
 
 type Messages = Record<string, string>;
 
@@ -271,8 +275,6 @@ const defaultMessages: Record<string, Messages> = {
   },
 };
 
-const I18nContext = createContext<I18nContextValue | undefined>(undefined);
-
 export function I18nProvider({ children }: { children: React.ReactNode }) {
   const [locale, setLocale] = useState<string>("en");
 
@@ -286,12 +288,6 @@ export function I18nProvider({ children }: { children: React.ReactNode }) {
   }), [locale]);
 
   return <I18nContext.Provider value={value}>{children}</I18nContext.Provider>;
-}
-
-export function useI18n() {
-  const ctx = useContext(I18nContext);
-  if (!ctx) throw new Error("useI18n must be used within I18nProvider");
-  return ctx;
 }
