+++ src/test-setup.d.ts
@@ -5,10 +5,10 @@ import 'vitest';
 import type { TestingLibraryMatchers } from '@testing-library/jest-dom/matchers';
 
 declare module 'vitest' {
-  interface Assertion<T = any> extends TestingLibraryMatchers<T, void> {
+  interface Assertion<T = unknown> extends TestingLibraryMatchers<T, void> {
     toHaveNoViolations(): T;
   }
-  interface AsymmetricMatchersContaining<T = any> extends TestingLibraryMatchers<T, void> {
+  interface AsymmetricMatchersContaining<T = unknown> extends TestingLibraryMatchers<T, void> {
     toHaveNoViolations(): T;
   }
 }
